utils: move get_seconds_since_j2000 onto FileHandler as a private method
The J2000 seconds scalar had exactly one consumer -- FileHandler's frame
clustering -- so it belongs with it, not in the shared kpf.py timestamp API.
Move it to FileHandler._seconds_since_j2000 and relocate its only dependency,
the _J2000_EPOCH constant, from kpf.py to io.py. kpf.py keeps
kpf_timestamp_to_datetime (still used by diagnostics/calibration_association),
which io.py now imports for the method body.

Tests move with the code: the coverage relocates from test_utils.py to a
TestSecondsSinceJ2000 class in test_masters_recipe.py, exercising the method
via a config-free FileHandler alongside the other FileHandler tests.
