+++ tests/regression/test_masters_recipe.py
@@ -171,6 +171,30 @@ def _cluster(cal_type, mini_db, **kwargs):
     return FileHandler().build_l0_file_lists(cal_type, mini_db=mini_db, **kwargs)
 
 
+class TestSecondsSinceJ2000:
+    """FileHandler._seconds_since_j2000 is the monotonic sort/gap scalar the
+    clustering builds on; exercised directly via a config-free handler."""
+
+    def test_basic(self):
+        # J2000.0 itself: 2000-01-01 12:00 UTC = '20000101.43200.00'
+        assert FileHandler()._seconds_since_j2000("20000101.43200.00") == 0
+
+    def test_monotonic_across_year_boundary(self):
+        # Dec 31 23:59:00 -> Jan 1 00:00:00 should differ by 60s exactly.
+        fh = FileHandler()
+        end = fh._seconds_since_j2000("20231231.86340.00")
+        start_next_year = fh._seconds_since_j2000("20240101.00000.00")
+        assert start_next_year - end == 60
+
+    def test_raises_on_invalid_timestamp(self):
+        with pytest.raises(ValueError, match="Invalid KPF timestamp"):
+            FileHandler()._seconds_since_j2000("KP.20240405.99999.57.fits")
+
+    def test_raises_when_no_timestamp_found(self):
+        with pytest.raises(ValueError, match="No KPF timestamp found"):
+            FileHandler()._seconds_since_j2000("notimestamp.fits")
+
+
 class TestBuildL0FileLists:
     """Clustering depends only on the mini database, so these exercise it with
     synthetic DataFrames (no files on disk) via the ``mini_db=`` override."""
